+++ packages/react-components/react-carousel-preview/library/src/components/CarouselContext.ts
@@ -4,10 +4,10 @@ import type { CarouselContextValue } from './CarouselContext.types';
 export const carouselContextDefaultValue: CarouselContextValue = {
   activeIndex: 0,
   selectPageByElement: () => {
-    /** noop */
+    return 0;
   },
   selectPageByDirection: () => {
-    /** noop */
+    return 0;
   },
   selectPageByIndex: () => {
     /** noop */