+++ packages/react-components/react-carousel-preview/library/src/components/CarouselNavButton/useCarouselNavButton.ts
@@ -1,4 +1,4 @@
-import { ARIAButtonElement, ARIAButtonSlotProps, useARIAButtonProps } from '@fluentui/react-aria';
+import { type ARIAButtonElement, type ARIAButtonSlotProps, useARIAButtonProps } from '@fluentui/react-aria';
 import { useTabsterAttributes } from '@fluentui/react-tabster';
 import {
   getIntrinsicElementProps,
@@ -12,6 +12,7 @@ import * as React from 'react';
 
 import { useCarouselContext_unstable as useCarouselContext } from '../CarouselContext';
 import { useCarouselNavContext } from '../CarouselNav/CarouselNavContext';
+import { useCarouselNavIndexContext } from '../CarouselNav/CarouselNavIndexContext';
 import type { CarouselNavButtonProps, CarouselNavButtonState } from './CarouselNavButton.types';
 
 /**
@@ -29,7 +30,9 @@ export const useCarouselNavButton_unstable = (
 ): CarouselNavButtonState => {
   const { onClick, as = 'button' } = props;
 
-  const { index, appearance } = useCarouselNavContext();
+  const { appearance } = useCarouselNavContext();
+  const index = useCarouselNavIndexContext();
+
   const selectPageByIndex = useCarouselContext(ctx => ctx.selectPageByIndex);
   const selected = useCarouselContext(ctx => ctx.activeIndex === index);
   const subscribeForValues = useCarouselContext(ctx => ctx.subscribeForValues);